Use a different algorithm to find anagrams faster
The current naive algorithm for finding anagrams takes forever (>5min).

Switch to a hash-based algorithm that finds anagrams by bucketting
words according to a normalized form.  This only works because we
are asked to find all anagrams in the entire file.

Performance is much better now:

    $ irb
    2.1.1 :001 > require './lib/anagram_finder'
     => true
    2.1.1 :002 > require 'benchmark'
     => true
    2.1.1 :003 > af = AnagramFinder.new(File.open("wordlist.txt"))
     => #<AnagramFinder:0x007ffaaedbbd60 @io=#<File:wordlist.txt>>
    2.1.1 :004 > Benchmark.realtime { puts af.anagrams.count }
    20683
     => 1.712718

And the code is actually shorter and simpler, too:
 1 file changed, 4 insertions(+), 20 deletions(-)