Optimize memory allocation by creating fewer arrays
The current implementation creates tons of arrays -- one extra one
for each word.

Adjust the "groups" hash to assign an empty array to each key accessed
so that we can easily push onto it instead of using "+=".

    $ irb
    2.1.1 :001 > require './lib/anagram_finder'
     => true
    2.1.1 :002 > require 'benchmark'
     => true
    2.1.1 :003 > af = AnagramFinder.new(File.open("wordlist.txt"))
     => #<AnagramFinder:0x007ff68e7f3c08 @io=#<File:wordlist.txt>>
    2.1.1 :004 > Benchmark.realtime { puts af.anagrams.count }
    20683
     => 1.640237

Speedup is minimal (4%), but we'll keep the change because the
readability is awash.  (This is not really a proper A/B performance
test; perhaps I should take more samples.)